+++ packages/docs-a/CHANGELOG.md
@@ -3,6 +3,62 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.5.1](https://github.com/6stars/gatsby-docs/compare/@m00n/docs-a@1.5.0...@m00n/docs-a@1.5.1) (2019-02-12)
+
+**Note:** Version bump only for package @m00n/docs-a
+
+# [1.5.0](https://github.com/6stars/gatsby-docs/compare/@m00n/docs-a@1.4.1...@m00n/docs-a@1.5.0) (2019-02-12)
+
+**Note:** Version bump only for package @m00n/docs-a
+
+## [1.4.1](https://github.com/6stars/gatsby-docs/compare/@m00n/docs-a@1.4.0...@m00n/docs-a@1.4.1) (2019-02-12)
+
+**Note:** Version bump only for package @m00n/docs-a
+
+# [1.4.0](https://github.com/6stars/gatsby-docs/compare/@m00n/docs-a@1.3.3...@m00n/docs-a@1.4.0) (2019-02-12)
+
+**Note:** Version bump only for package @m00n/docs-a
+
+## [1.3.3](https://github.com/6stars/gatsby-docs/compare/@m00n/docs-a@1.3.2...@m00n/docs-a@1.3.3) (2019-02-12)
+
+**Note:** Version bump only for package @m00n/docs-a
+
+## [1.3.2](https://github.com/6stars/gatsby-docs/compare/@m00n/docs-a@1.3.1...@m00n/docs-a@1.3.2) (2019-02-12)
+
+**Note:** Version bump only for package @m00n/docs-a
+
+## [1.3.1](https://github.com/6stars/gatsby-docs/compare/@m00n/docs-a@1.3.0...@m00n/docs-a@1.3.1) (2019-02-12)
+
+**Note:** Version bump only for package @m00n/docs-a
+
+# [1.3.0](https://github.com/6stars/gatsby-docs/compare/@m00n/docs-a@1.2.1...@m00n/docs-a@1.3.0) (2019-02-12)
+
+**Note:** Version bump only for package @m00n/docs-a
+
+## [1.2.1](https://github.com/6stars/gatsby-docs/compare/@m00n/docs-a@1.2.0...@m00n/docs-a@1.2.1) (2019-02-12)
+
+**Note:** Version bump only for package @m00n/docs-a
+
+# [1.2.0](https://github.com/6stars/gatsby-docs/compare/@m00n/docs-a@1.1.0...@m00n/docs-a@1.2.0) (2019-02-12)
+
+**Note:** Version bump only for package @m00n/docs-a
+
+# [1.1.0](https://github.com/6stars/gatsby-docs/compare/@m00n/docs-a@1.0.6...@m00n/docs-a@1.1.0) (2019-02-12)
+
+**Note:** Version bump only for package @m00n/docs-a
+
+## [1.0.6](https://github.com/6stars/gatsby-docs/compare/@m00n/docs-a@1.0.5...@m00n/docs-a@1.0.6) (2019-02-12)
+
+**Note:** Version bump only for package @m00n/docs-a
+
+## [1.0.5](https://github.com/6stars/gatsby-docs/compare/@m00n/docs-a@1.0.4...@m00n/docs-a@1.0.5) (2019-02-11)
+
+**Note:** Version bump only for package @m00n/docs-a
+
+## [1.0.4](https://github.com/6stars/gatsby-docs/compare/@m00n/docs-a@1.0.3...@m00n/docs-a@1.0.4) (2019-02-11)
+
+**Note:** Version bump only for package @m00n/docs-a
+
 ## [1.0.3](https://github.com/6stars/gatsby-docs/compare/@m00n/docs-a@1.0.2...@m00n/docs-a@1.0.3) (2019-02-11)
 
 **Note:** Version bump only for package @m00n/docs-a