+++ packages/gatsby-docs-ui/CHANGELOG.md
@@ -3,6 +3,68 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.6.1](https://github.com/6stars/gatsby-docs/compare/@m00n/gatsby-docs-ui@1.6.0...@m00n/gatsby-docs-ui@1.6.1) (2019-02-12)
+
+**Note:** Version bump only for package @m00n/gatsby-docs-ui
+
+# [1.6.0](https://github.com/6stars/gatsby-docs/compare/@m00n/gatsby-docs-ui@1.5.1...@m00n/gatsby-docs-ui@1.6.0) (2019-02-12)
+
+**Note:** Version bump only for package @m00n/gatsby-docs-ui
+
+## [1.5.1](https://github.com/6stars/gatsby-docs/compare/@m00n/gatsby-docs-ui@1.5.1...@m00n/gatsby-docs-ui@1.5.1) (2019-02-12)
+
+**Note:** Version bump only for package @m00n/gatsby-docs-ui
+
+## [1.5.1](https://github.com/6stars/gatsby-docs/compare/@m00n/gatsby-docs-ui@1.5.0...@m00n/gatsby-docs-ui@1.5.1) (2019-02-12)
+
+**Note:** Version bump only for package @m00n/gatsby-docs-ui
+
+# [1.5.0](https://github.com/6stars/gatsby-docs/compare/@m00n/gatsby-docs-ui@1.4.3...@m00n/gatsby-docs-ui@1.5.0) (2019-02-12)
+
+**Note:** Version bump only for package @m00n/gatsby-docs-ui
+
+## [1.4.3](https://github.com/6stars/gatsby-docs/compare/@m00n/gatsby-docs-ui@1.4.2...@m00n/gatsby-docs-ui@1.4.3) (2019-02-12)
+
+**Note:** Version bump only for package @m00n/gatsby-docs-ui
+
+## [1.4.2](https://github.com/6stars/gatsby-docs/compare/@m00n/gatsby-docs-ui@1.4.1...@m00n/gatsby-docs-ui@1.4.2) (2019-02-12)
+
+**Note:** Version bump only for package @m00n/gatsby-docs-ui
+
+## [1.4.1](https://github.com/6stars/gatsby-docs/compare/@m00n/gatsby-docs-ui@1.3.1...@m00n/gatsby-docs-ui@1.4.1) (2019-02-12)
+
+**Note:** Version bump only for package @m00n/gatsby-docs-ui
+
+# 1.4.0 (2019-02-12)
+
+### Bug Fixes
+
+- module-resolver babel ([13ac272](https://github.com/6stars/gatsby-docs/commit/13ac272))
+
+## [1.3.1](https://github.com/6stars/gatsby-docs/compare/@m00n/gatsby-docs-ui@1.3.0...@m00n/gatsby-docs-ui@1.3.1) (2019-02-12)
+
+**Note:** Version bump only for package @m00n/gatsby-docs-ui
+
+# [1.3.0](https://github.com/6stars/gatsby-docs/compare/@m00n/gatsby-docs-ui@1.2.0...@m00n/gatsby-docs-ui@1.3.0) (2019-02-12)
+
+**Note:** Version bump only for package @m00n/gatsby-docs-ui
+
+# [1.2.0](https://github.com/6stars/gatsby-docs/compare/@m00n/gatsby-docs-ui@1.1.6...@m00n/gatsby-docs-ui@1.2.0) (2019-02-12)
+
+**Note:** Version bump only for package @m00n/gatsby-docs-ui
+
+## [1.1.6](https://github.com/6stars/gatsby-docs/compare/@m00n/gatsby-docs-ui@1.1.5...@m00n/gatsby-docs-ui@1.1.6) (2019-02-12)
+
+**Note:** Version bump only for package @m00n/gatsby-docs-ui
+
+## [1.1.5](https://github.com/6stars/gatsby-docs/compare/@m00n/gatsby-docs-ui@1.1.4...@m00n/gatsby-docs-ui@1.1.5) (2019-02-11)
+
+**Note:** Version bump only for package @m00n/gatsby-docs-ui
+
+## [1.1.4](https://github.com/6stars/gatsby-docs/compare/@m00n/gatsby-docs-ui@1.1.3...@m00n/gatsby-docs-ui@1.1.4) (2019-02-11)
+
+**Note:** Version bump only for package @m00n/gatsby-docs-ui
+
 ## [1.1.3](https://github.com/6stars/gatsby-docs/compare/@m00n/gatsby-docs-ui@1.1.2...@m00n/gatsby-docs-ui@1.1.3) (2019-02-11)
 
 ### Bug Fixes