+++ packages/gatsby-plugin-nav-tree/CHANGELOG.md
@@ -3,6 +3,34 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.6.0](https://github.com/6stars/gatsby-docs/compare/@m00n/gatsby-plugin-nav-tree@0.5.1...@m00n/gatsby-plugin-nav-tree@0.6.0) (2019-02-12)
+
+**Note:** Version bump only for package @m00n/gatsby-plugin-nav-tree
+
+## [0.5.1](https://github.com/6stars/gatsby-docs/compare/@m00n/gatsby-plugin-nav-tree@0.5.0...@m00n/gatsby-plugin-nav-tree@0.5.1) (2019-02-12)
+
+**Note:** Version bump only for package @m00n/gatsby-plugin-nav-tree
+
+# [0.5.0](https://github.com/6stars/gatsby-docs/compare/@m00n/gatsby-plugin-nav-tree@0.4.0...@m00n/gatsby-plugin-nav-tree@0.5.0) (2019-02-12)
+
+**Note:** Version bump only for package @m00n/gatsby-plugin-nav-tree
+
+# [0.4.0](https://github.com/6stars/gatsby-docs/compare/@m00n/gatsby-plugin-nav-tree@0.3.6...@m00n/gatsby-plugin-nav-tree@0.4.0) (2019-02-12)
+
+**Note:** Version bump only for package @m00n/gatsby-plugin-nav-tree
+
+## [0.3.6](https://github.com/6stars/gatsby-docs/compare/@m00n/gatsby-plugin-nav-tree@0.3.5...@m00n/gatsby-plugin-nav-tree@0.3.6) (2019-02-12)
+
+**Note:** Version bump only for package @m00n/gatsby-plugin-nav-tree
+
+## [0.3.5](https://github.com/6stars/gatsby-docs/compare/@m00n/gatsby-plugin-nav-tree@0.3.4...@m00n/gatsby-plugin-nav-tree@0.3.5) (2019-02-11)
+
+**Note:** Version bump only for package @m00n/gatsby-plugin-nav-tree
+
+## [0.3.4](https://github.com/6stars/gatsby-docs/compare/@m00n/gatsby-plugin-nav-tree@0.3.3...@m00n/gatsby-plugin-nav-tree@0.3.4) (2019-02-11)
+
+**Note:** Version bump only for package @m00n/gatsby-plugin-nav-tree
+
 ## [0.3.3](https://github.com/6stars/gatsby-docs/compare/@m00n/gatsby-plugin-nav-tree@0.3.2...@m00n/gatsby-plugin-nav-tree@0.3.3) (2019-02-11)
 
 **Note:** Version bump only for package @m00n/gatsby-plugin-nav-tree